+++ src/netlink_pm.c
@@ -15,27 +15,10 @@
 
 #include "netlink_pm.h"
 
-/// Directory containing MPTCP sysctl variable entries.
-#define MPTCP_SYSCTL_BASE "/proc/sys/net/mptcp/"
 
-/// Constuct full path for MPTCP sysctl variable "@a name".
-#define MPTCP_SYSCTL_VARIABLE(name) MPTCP_SYSCTL_BASE #name
-
-/// Get multipath-tcp.org kernel generic netlink PM characteristics.
-struct mptcpd_netlink_pm const *mptcpd_get_netlink_pm_mptcp_org(void);
-
-/// Get upstream kernel generic netlink PM characteristics.
-struct mptcpd_netlink_pm const *mptcpd_get_netlink_pm_upstream(void);
-
-/**
- * @brief Verify that MPTCP is enabled at run-time in the kernel.
- *
- * Check that MPTCP is enabled through the specified @c sysctl
- * variable.
- */
-static bool check_kernel_mptcp_enabled(char const *path,
-                                       char const *variable,
-                                       int enable_val)
+bool mptcpd_is_kernel_mptcp_enabled(char const *path,
+                                    char const *variable,
+                                    int enable_val)
 {
         FILE *const f = fopen(path, "r");
 
@@ -70,38 +53,6 @@ static bool check_kernel_mptcp_enabled(char const *path,
 
         return enabled;
 }
-
-/// Are we being run under a MPTCP-capable upstream kernel?
-static bool is_upstream_kernel(void)
-{
-        static char const path[] = MPTCP_SYSCTL_VARIABLE(enabled);
-        static char const name[] = "enabled";
-        static int  const enable_val = 1;
-
-        return check_kernel_mptcp_enabled(path, name, enable_val);
-}
-
-/// Are we being run under a MPTCP-capable multipath-tcp.org kernel?
-static bool is_mptcp_org_kernel(void)
-{
-        static char const path[] = MPTCP_SYSCTL_VARIABLE(mptcp_enabled);
-        static char const name[] = "mptcp_enabled";
-        static int  const enable_val = 2;  // or 1
-
-        return check_kernel_mptcp_enabled(path, name, enable_val);
-}
-
-struct mptcpd_netlink_pm const *mptcpd_get_netlink_pm(void)
-{
-        if (is_upstream_kernel()) {
-                return mptcpd_get_netlink_pm_upstream();
-        } else if (is_mptcp_org_kernel()) {
-                return mptcpd_get_netlink_pm_mptcp_org();
-        } else {
-                return NULL;
-        }
-}
-
 
 /*
   Local Variables:
